Smoosh everything into one project
