+++ src/torch/shared/index.js
