+++ src/torch/shared/js/uploadr.js
