+++ src/torch/shared/styles/upload.css
