--- src/torchweb/requirements.txt
@@ -1,9 +1,0 @@
-alembic==1.7.7
-Flask==2.1.2
-Flask_Mail==0.9.1
-Flask_Migrate==3.1.0
-flask_security==3.0.0
-Flask_Security_Too==4.1.4
-Flask_SQLAlchemy==2.5.1
-SQLAlchemy==1.4.36
-WTForms==3.0.1
